huge commit COM + WIN32 + Project

+++ win32/Practice/DLLForProject/PhysicsDll/PhysicsDll.h
@@ -1,1 +1,1 @@
-extern "C" int WeightCalculator(int, int);
+extern "C" double WeightCalculator(double, int);
